+++ CHANGELOG.md
@@ -11,6 +11,8 @@ All notable changes to this project will be documented in this file.
 - Customer Detail Page with Sign-Out feaature
 - Validator Login form[@phillip](https://github.com/hemvanh/TPTEA/pull/23)
 - Routing with new token of Admin & Customer[@phillip](https://github.com/hemvanh/TPTEA/pull/24)
+- Test case routing (No-Token, TokenAdmin, TokenCustomer)
+- Automation Test Routing (Testcafe framework)
 ### Removed 
 - Join now button
 ### Fixed